refactor(main): group kubernetes logic
