+++ packages/main/src/plugin/kubernetes/kubernetes-client.ts
@@ -67,19 +67,19 @@ import type * as containerDesktopAPI from '@podman-desktop/api';
 import * as jsYaml from 'js-yaml';
 import { parseAllDocuments } from 'yaml';
 
-import type { Telemetry } from '/@/plugin/telemetry/telemetry.js';
 import type { KubeContext } from '/@api/kubernetes-context.js';
 import type { V1Route } from '/@api/openshift-types.js';
 
-import type { ApiSenderType } from './api.js';
-import type { PodInfo } from './api/pod-info.js';
-import type { ConfigurationRegistry, IConfigurationNode } from './configuration-registry.js';
-import { Emitter } from './events/emitter.js';
-import type { FilesystemMonitoring } from './filesystem-monitoring.js';
+import type { ApiSenderType } from '../api.js';
+import type { PodInfo } from '../api/pod-info.js';
+import type { ConfigurationRegistry, IConfigurationNode } from '../configuration-registry.js';
+import { Emitter } from '../events/emitter.js';
+import type { FilesystemMonitoring } from '../filesystem-monitoring.js';
+import type { Telemetry } from '../telemetry/telemetry.js';
+import { Uri } from '../types/uri.js';
 import type { ContextGeneralState, ResourceName } from './kubernetes-context-state.js';
 import { ContextsManager } from './kubernetes-context-state.js';
 import { BufferedStreamWriter, ResizableTerminalWriter, StringLineReader } from './kubernetes-exec-transmitter.js';
-import { Uri } from './types/uri.js';
 
 interface KubernetesObjectWithKind extends KubernetesObject {
   kind: string;
