+++ packages/main/src/plugin/kubernetes/kubernetes-context-state-constants.ts
