+++ packages/renderer/src/lib/node/NodeEmptyScreen.svelte
@@ -3,7 +3,7 @@ import { EmptyScreen } from '@podman-desktop/ui-svelte';
 
 import { kubernetesCurrentContextState } from '/@/stores/kubernetes-contexts-state';
 
-import type { ContextGeneralState } from '../../../../main/src/plugin/kubernetes-context-state';
+import type { ContextGeneralState } from '../../../../main/src/plugin/kubernetes/kubernetes-context-state';
 import NodeIcon from '../images/NodeIcon.svelte';
 
 // If the current context is CONNECTED and we are on this empty screen
