+++ packages/renderer/src/lib/preferences/PreferencesKubernetesContextsRendering.spec.ts
@@ -26,7 +26,10 @@ import { kubernetesContexts } from '/@/stores/kubernetes-contexts';
 import * as kubernetesContextsState from '/@/stores/kubernetes-contexts-state';
 import type { KubeContext } from '/@api/kubernetes-context';
 
-import type { CheckingState, ContextGeneralState } from '../../../../main/src/plugin/kubernetes-context-state';
+import type {
+  CheckingState,
+  ContextGeneralState,
+} from '../../../../main/src/plugin/kubernetes/kubernetes-context-state';
 import PreferencesKubernetesContextsRendering from './PreferencesKubernetesContextsRendering.svelte';
 
 vi.mock('/@/stores/kubernetes-contexts-state', async () => {
