+++ packages/renderer/src/lib/ui/KubernetesCurrentContextConnectionBadge.spec.ts
@@ -25,7 +25,7 @@ import { beforeEach, expect, test, vi } from 'vitest';
 
 import KubernetesCurrentContextConnectionBadge from '/@/lib/ui/KubernetesCurrentContextConnectionBadge.svelte';
 
-import type { ContextGeneralState } from '../../../../main/src/plugin/kubernetes-context-state';
+import type { ContextGeneralState } from '../../../../main/src/plugin/kubernetes/kubernetes-context-state';
 
 const mocks = vi.hoisted(() => ({
   subscribeMock: vi.fn(),
